+++ core/src/runtime/mod.rs
@@ -1,0 +1,3 @@
+mod eval;
+
+pub(crate) use eval::{eval_expr, eval_query};
