+++ core/tests/spans.rs
@@ -27,7 +27,7 @@ fn test(input: &str, output: &str) {
     let mut iter = text_query::TokenIterator::new(input.trim()).peekable();
     let expr = text_query::parse_query(&mut iter);
     CONTEXT.with(|ctx| {
-        let res = ctx.eval_outer(&expr);
+        let res = ctx.eval_query(&expr);
         let mut string = String::new();
         match res {
             Ok(v) => write_string(&mut string, &v),
